app.js separated for testing-first riuote tested - receiving 404" g"
