+++ package.json
@@ -34,6 +34,7 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.1",
-    "@testing-library/react": "^12.1.2"
+    "@testing-library/react": "^12.1.2",
+    "supertest": "^6.1.6"
   }
 }
