+++ server.js
@@ -6,29 +6,29 @@ const cookieParser = require('cookie-parser')
 const session = require('express-session')
 const flash = require('connect-flash')
 const express = require('express');
-const app = express();
+const app = require('./app');
 
 // seed database
 const seed = require('./seed');
 seed();
 
-// send data as json object
-app.use(express.urlencoded({ extended: true }));
-app.use(express.json());
+// // send data as json object
+// app.use(express.urlencoded({ extended: true }));
+// app.use(express.json());
 
-// enable handlebars frontend
-app.engine(
-  'handlebars',
-  engine({
-    runtimeOptions: {
-      allowProtoPropertiesByDefault: true,
-      allowProtoMethodsByDefault: true,
-    },
-  })
-);
+// // enable handlebars frontend
+// app.engine(
+//   'handlebars',
+//   engine({
+//     runtimeOptions: {
+//       allowProtoPropertiesByDefault: true,
+//       allowProtoMethodsByDefault: true,
+//     },
+//   })
+// );
 
-app.set('view engine', 'handlebars');
-app.use(express.static('public'));
+// app.set('view engine', 'handlebars');
+// app.use(express.static('public'));
 
 // custom handlebar operator helpers
 // pulled from: https://stackoverflow.com/questions/33316562/how-to-compare-a-value-in-handlebars
